refactor character further
